% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_qr.R
\name{barcode_qr}
\alias{barcode_qr}
\title{Barcode generator}
\usage{
barcode_qr(text, color = "black", alpha = 1)
}
\arguments{
\item{text}{text to convert to QR bar code}

\item{color}{Bar code color}

\item{alpha}{Intensity of the bar code color}
}
\value{
plot
}
\description{
Generate bar codes using QR codes
}
\examples{

\dontrun{

barcode_qr("inkaverse")

}

}
