% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huito_fonts.R
\name{huito_fonts}
\alias{huito_fonts}
\title{Fonts import}
\usage{
huito_fonts(fonts = NA)
}
\arguments{
\item{fonts}{fonts names. Default "Permanent Marker"}
}
\value{
fonts
}
\description{
Import fonts from Google fonts
}
\details{
For more fonts visit: https://fonts.google.com/
}
