% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huito.R
\name{huito_labels}
\alias{huito_labels}
\title{Label generator}
\usage{
huito_labels(
  data,
  options,
  ncol = 2,
  complete = FALSE,
  file = "labels.pdf",
  dpi = 100
)
}
\arguments{
\item{data}{data frame to build the labels}

\item{options}{data frame. See details}

\item{ncol}{Number of column for the label in complete mode}

\item{complete}{logical value for build all labels}

\item{file}{path of the final file}

\item{dpi}{images resolution}
}
\value{
image or pdf
}
\description{
Generate labels based in a data frame
}
\details{
`options` is build with the function: `huito_label_options()`

The Labels are build by layers

`elements` = "label" | "barcode" | "figure" | "text"
`type` = "template" | "column" | "manual"
`value` = 'column' name from data frame or 'manual' for write personal info
`dimension` = dimension for the labels and images
`position` = values represented by the unit used for position(x, y) inside the label
`color` = color by the name or code
`margin` = space between labels
`unit` = units used for the design and image dimension
`border` = color border for the label
`size` = font size
`angle` = angle for the text
`font` = font type from Google fonts

 For more details, see example
}
\examples{

\dontrun{

library(gsheet)

url_fb <- paste0("https://docs.google.com/spreadsheets/d/"
         , "1q0EZmZBt52ca-0VbididjJy2jXTwf06laJpzvkQJWvc/edit#gid=107939497")

fb <- gsheet2tbl(url_fb)

lyr_temp <- list(element = "label"
                 , type = "template"
                 , dimension = "10*2.5"
                 , color = "white"
                 , margin = 0.04
                 , unit = "cm"
                 , border = "red")

lyr_qr <- list(element = "barcode"
               , type = "column"
               , value = "qr-code"
               , position = "7*0.1"
               , dimension = "2.3*2.3"
               , unit = "cm")

lyr_pic <- list(element = "figure"
                , type = "column"
                , value = "image"
                , position = "2.3*0.5"
                , dimension = "1.5*1.5"
                , unit = "cm")

lyr_wf <- list(element = "figure"
               , type = "manual"
               , value = paste0("https://upload.wikimedia.org/"
                                , "wikipedia/commons/a/a9/Unalm_logo.png")
               , position = "0.1*0.2"
               , dimension = "2*2"
               , unit = "cm")

txt5 <- list(element = "text"
             , type = "column"
             , value = "variedad"
             , size = 11
             , angle = 0
             , font = "Permanent Marker"
             , color = "black"
             , position = "5.5*1.6")

txt6 <- list(element = "text"
             , type = "column"
             , value = "temperatura"
             , size = 11
             , angle = 0
             , font = "Permanent Marker"
             , color = "black"
             , position = "5.5*0.5")

txt7 <- list(element = "text"
             , type = "column"
             , value = "plots"
             , size = 11
             , angle = 90
             , font = "Permanent Marker"
             , color = "blue"
             , position = "9.7*1.3")

txt8 <- list(element = "text"
             , type = "manual"
             , value = "Variedad"
             , size = 12
             , angle = 0
             , font = "Permanent Marker"
             , color = "red"
             , position = "5.5*2.1")

txt9 <- list(element = "text"
             , type = "manual"
             , value = "Temperature"
             , size = 12
             , angle = 0
             , font = "Permanent Marker"
             , color = "red"
             , position = "5.5*1")

lyr_bg <- list(element = "figure"
               , type = "manual"
               , value = paste0("https://imborrable.com/wp-content/"
                    , "uploads/2021/01/psicologia-del-color-significados.png")
               , position = "0*0"
               , dimension = "10*5"
               , unit = "cm")

rm(lyr_bg)

labels <- huito_label_options(c("lyr", "txt")) \%>\%
  huito_labels(data = fb, options = ., complete = F)

}

}
