% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_barcode.R
\name{include_barcode}
\alias{include_barcode}
\title{Barcode layer}
\usage{
include_barcode(
  label,
  value,
  size,
  position = NA,
  type = "static",
  color = "auto",
  units = "cm"
)
}
\arguments{
\item{label}{label output}

\item{value}{column or path}

\item{size}{image size}

\item{position}{position coordinate}

\item{type}{type of entry: dynamic or static}

\item{color}{image color}

\item{units}{units for the label options}
}
\value{
data frame
}
\description{
Insert barcode in label
}
\examples{

\dontrun{

library(gsheet)

url <- paste0("https://docs.google.com/spreadsheets/d/"
       , "1q0EZmZBt52ca-0VbididjJy2jXTwf06laJpzvkQJWvc/edit#gid=107939497")
fb <- gsheet2tbl(url)

label <- label(data = fb
               , size = c(10, 2.5)
               ) \%>\%
         include_image(
               value = "https://inkaverse.com/reference/figures/logo.png"
               , size = c(2,2)
               ) \%>\%
         include_barcode(
               value = "qr-code"
               , size = c(2,2)
               )

}

}
