% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_text.R
\name{include_text}
\alias{include_text}
\title{Text layer}
\usage{
include_text(
  label,
  value,
  prefix = "",
  position = NA,
  fontface = "plain",
  size = 11,
  font = NA,
  type = "static",
  color = NA,
  angle = 0,
  opts = NA
)
}
\arguments{
\item{label}{label output}

\item{value}{column or string}

\item{prefix}{Prefix for the dynamic text}

\item{position}{position coordinate}

\item{fontface}{font type \verb{[character: plain, bold, italic]}}

\item{size}{text size}

\item{font}{font type}

\item{type}{type of entry \verb{[character: dynamic or static]}}

\item{color}{image color}

\item{angle}{angle of the text}

\item{opts}{list arguments from draw_label()}
}
\value{
data frame
}
\description{
Insert text in label
}
