% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Label layout}
\usage{
label(
  data,
  size,
  background = NA,
  border_color = NA,
  border_width = NA,
  units = "cm"
)
}
\arguments{
\item{data}{data frame to build the labels}

\item{size}{label size}

\item{background}{background color}

\item{border_color}{border color}

\item{border_width}{border width}

\item{units}{units for the label options}

\item{margin}{margin size}

\item{paper}{paper size}

\item{dpi}{label resolution}
}
\value{
data frame
}
\description{
Generate labels options
}
\examples{

\dontrun{

library(gsheet)

url <- "https://docs.google.com/spreadsheets/d/1q0EZmZBt52ca-0VbididjJy2jXTwf06laJpzvkQJWvc/edit#gid=107939497"
fb <- gsheet2tbl(url)

label <- label(data = fb
               , size = c(10, 2.5)
               )

}

}
