% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_hexagon.R
\name{shape_hexagon}
\alias{shape_hexagon}
\title{Shape hexagon}
\usage{
shape_hexagon(
  border_width = 2,
  background = NA,
  border_color = "black",
  margin = 0,
  units = "cm",
  transparent = TRUE
)
}
\arguments{
\item{border_width}{line width}

\item{background}{background color}

\item{border_color}{border color}

\item{margin}{margin}

\item{units}{units for shape}

\item{transparent}{transparent borders}
}
\value{
geom
}
\description{
Hexagon geom shape for ggplot2
}
\examples{

shape_hexagon(border_width = 5, background = "red",  border_color = "blue")

}
