% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_qr.R
\name{barcode_qr}
\alias{barcode_qr}
\title{Barcode generator}
\usage{
barcode_qr(text, color = "black", alpha = 1, ecl = "H")
}
\arguments{
\item{text}{text to convert to QR bar code}

\item{color}{Bar code color}

\item{alpha}{Intensity of the bar code color}

\item{ecl}{Error correction level (percentage). "L" (7), "M" (15), "Q" (25)
and "H" (30). Defaults to "H"}
}
\value{
plot
}
\description{
Generate bar codes using QR codes
}
\examples{

library(huito)

barcode_qr("LIMA-2021-11-03_15_3_4")


}
