% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldbook.R
\docType{data}
\name{fieldbook}
\alias{fieldbook}
\title{Fieldbook experimental design}
\format{
A data frame with 24 rows and 5 variables:

\describe{

\item{barcode}{barcode for each experimental unit}
\item{plots}{Plot number}
\item{block}{Blocks (3): number of replication in the design}
\item{condition}{Factor with two levels: irrigated and drought}
\item{genotypes}{Factor with four levels: choclito, salcedo, pandela, puno}

}
}
\usage{
fieldbook
}
\description{
The dataset were obtained using inti package. The data set is a randomize
complete block design (RCBD) with three replications.
}
\keyword{datasets}
