% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{image_import}
\alias{image_import}
\title{Image import}
\usage{
image_import(image, opts = NA)
}
\arguments{
\item{image}{image path or URL}

\item{opts}{R magick functions by layers}
}
\value{
image
}
\description{
Import images and apply R magick functions
}
