% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_import.R
\name{image_import}
\alias{image_import}
\title{Image import options}
\usage{
image_import(image, opts = NA)
}
\arguments{
\item{image}{path or url}

\item{opts}{R magick options by layers}
}
\value{
image
}
\description{
Import images and include R magick options
}
