% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_barcode.R
\name{include_barcode}
\alias{include_barcode}
\title{Barcode layer}
\usage{
include_barcode(
  label,
  value,
  size,
  position = NA,
  type = "static",
  color = "auto",
  units = "cm"
)
}
\arguments{
\item{label}{label output}

\item{value}{column or path}

\item{size}{image size}

\item{position}{position coordinate}

\item{type}{type of entry: dynamic or static}

\item{color}{image color}

\item{units}{units for the label options}
}
\value{
data frame
}
\description{
Insert barcode in label
}
