% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_image.R
\name{include_image}
\alias{include_image}
\title{Image layer}
\usage{
include_image(
  label,
  value,
  size,
  position = NA,
  type = "static",
  units = "cm",
  opts = NA
)
}
\arguments{
\item{label}{label output}

\item{value}{column or path}

\item{size}{image size}

\item{position}{position coordinate}

\item{type}{type of entry: dynamic or static}

\item{units}{units for the label options}

\item{opts}{R magick functions}
}
\value{
data frame
}
\description{
Insert image in label
}
\examples{

library(huito)

label <- label_layout(size = c(10, 2.5), border_color = "blue") \%>\%
  include_image(
    value = "https://flavjack.github.io/inti/img/inkaverse.png"
    , size = c(2.4, 2.4)
    , position = c(1.2, 1.25)
    )

}
