% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_shape.R
\name{include_shape}
\alias{include_shape}
\title{Shape layer}
\usage{
include_shape(
  label,
  value = "hexagon",
  size = NA,
  position = NA,
  border_color = "black",
  border_width = 2,
  background = NA,
  margin = 0,
  units = "cm"
)
}
\arguments{
\item{label}{label output}

\item{value}{column or path}

\item{size}{image size}

\item{position}{position coordinate}

\item{border_color}{image color}

\item{border_width}{shape line width}

\item{background}{background color}

\item{margin}{= shape margin}

\item{units}{units for shape}
}
\value{
data frame
}
\description{
Insert shape in label
}
\examples{

\dontrun{

label <- label_layout(data = NA
               , size = c(10, 2.5)
               ) \%>\%
         include_shape(
               value = "hexagon"
               ) 

}

}
