% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_layout.R
\name{label_layout}
\alias{label_layout}
\title{Label layout}
\usage{
label_layout(
  data = NA,
  size,
  background = "white",
  border_color = "black",
  border_width = 0.5,
  units = "cm"
)
}
\arguments{
\item{data}{data frame to build the labels}

\item{size}{label size}

\item{background}{background color}

\item{border_color}{border color}

\item{border_width}{border width}

\item{units}{units for the label options}
}
\value{
data frame
}
\description{
Generate labels options
}
\examples{

label <- label_layout(size = c(10, 2.5)
                   , border_color = "red"
                   , border_width = 1
                   ) \%>\% 
                 label_print()

}
