% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_print.R
\name{label_print}
\alias{label_print}
\title{Label print}
\usage{
label_print(
  label,
  filename = "label",
  mode = "sample",
  margin = NA,
  paper = NA,
  units = "cm",
  dpi = 100
)
}
\arguments{
\item{label}{data frame to build the labels}

\item{filename}{path of the final file}

\item{mode}{label in sample or complete mode}

\item{dpi}{images resolution}
}
\value{
png or pdf
}
\description{
Generate labels based in a data frame
}
\examples{

\dontrun{

library(gsheet)

url <- "https://docs.google.com/spreadsheets/d/1q0EZmZBt52ca-0VbididjJy2jXTwf06laJpzvkQJWvc/edit#gid=107939497"
fb <- gsheet2tbl(url)

label <- label(data = fb
               , size = c(10, 2.5)
               ) \%>\%
         include_image(
               value = "https://inkaverse.com/reference/figures/logo.png"
               , size = c(2.2,2.2)
               , position = c(0.2, 0.15)
               ) \%>\%
         include_barcode(
               value = "qr-code"
               , size = c(2.2, 2.2)
               , position = c(7, 0.15)
               ) \%>\%
         include_text(value = "plots", position = c(9.7, 1.25), angle = 90) \%>\%
         label_print()

}

}
