% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_print.R
\name{label_print}
\alias{label_print}
\title{Label print}
\usage{
label_print(
  label,
  mode = "sample",
  filename = "labels",
  margin = 0.04,
  paper = c(21, 29.7),
  units = "cm",
  viewer = FALSE,
  smpres = 200
)
}
\arguments{
\item{label}{data frame to build the labels}

\item{mode}{label in "sample" or "complete" mode}

\item{filename}{labels file name}

\item{margin}{labels margins. margin(t = 0, r = 0, b = 0, l = 0)}

\item{paper}{paper size}

\item{units}{units for the label options}

\item{viewer}{show the sample in the "Plots" or "Viewer" panel}

\item{smpres}{sample resolution}
}
\value{
pdf
}
\description{
Generate labels based in a data frame
}
\examples{

\dontrun{

library(huito)
library(gsheet)

url <- paste0("https://docs.google.com/spreadsheets/d/"
       , "1q0EZmZBt52ca-0VbididjJy2jXTwf06laJpzvkQJWvc/edit#gid=107939497")
fb <- gsheet2tbl(url)

label <- label_layout(data = fb
                     , size = c(10, 2.5)
                     ) \%>\%
         include_image(
               value = "https://flavjack.github.io/inti/img/inkaverse.png"
               , size = c(2.4, 2.4)
               , position = c(0.1, 0.05)
               ) \%>\%
         include_barcode(
               value = "qr-code"
               , size = c(2.2, 2.2)
               , position = c(7.2, 0.15)
               ) \%>\%
         include_text(value = "plots", position = c(9.7, 1.25), angle = 90, size = 15) \%>\%
         include_text(value = "Inkaverse", position = c(4.5, 1.25), size = 18) \%>\%
         label_print("sample")

}

}
