% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_hexagon.R
\name{shape_hexagon}
\alias{shape_hexagon}
\title{Shape hexagon}
\usage{
shape_hexagon(
  size = 5.08,
  border_width = 1,
  background = "transparent",
  border_color = "black",
  units = "cm",
  panel_color = "green",
  panel_size = NA
)
}
\arguments{
\item{size}{hexagon size (numeric: 5.08)}

\item{border_width}{line width (numeric: 1)}

\item{background}{background color (string: "transparent")}

\item{border_color}{border color (string: "black")}

\item{units}{units for shape (string: "cm")}

\item{panel_color}{panel color (string: "green")}

\item{panel_size}{panel size (numeric: NA)}
}
\value{
geom
}
\description{
Hexagon geom shape for ggplot2
}
\examples{

library(huito)

shape_hexagon(border_width = 1
              , background = "red"
              #, panel_size = 5.08
              )

}
