% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_hexagon.R
\name{shape_hexagon}
\alias{shape_hexagon}
\title{Shape hexagon}
\usage{
shape_hexagon(
  size = 4.39,
  border_width = 2,
  background = "transparent",
  border_color = "black",
  margin = 0,
  units = "cm",
  panel_color = "green",
  panel_size = 5.08
)
}
\arguments{
\item{size}{hexagon size}

\item{border_width}{line width}

\item{background}{background color}

\item{border_color}{border color}

\item{margin}{margin}

\item{units}{units for shape}

\item{panel_color}{panel color}

\item{panel_size}{panel size}
}
\value{
geom
}
\description{
Hexagon geom shape for ggplot2
}
\examples{

shape_hexagon(border_width = 3)

}
